+++ crates/krb5-kdc/src/lib.rs
@@ -88,8 +88,9 @@ pub use store::{
     KDB_PWCHANGE_SERVICE, KDB_REQUIRES_HW_AUTH, KDB_REQUIRES_PRE_AUTH, KDB_REQUIRES_PWCHANGE,
     KDB_V1_BASE_LENGTH, KadmData, KeyEntry, KeyLookup, MAX_ALIAS_DEPTH, NamedPolicy, PWQUAL_DICT,
     PWQUAL_EMPTY, PWQUAL_PRINC, Policy, Principal, PrincipalStore, RID_ADMINISTRATOR,
-    RID_FIRST_USER, RID_KRBTGT, S2K_ITERS, TlData, UlogEntry, db_args_put_error, kadm5_mask,
-    parse_dict_words, parse_spake_preauth_groups, random_key, s2k_params, strip_db_args,
+    RID_FIRST_USER, RID_KRBTGT, S2K_ITERS, TlData, UlogEntry, apply_keysalt_policy,
+    db_args_put_error, kadm5_mask, parse_dict_words, parse_spake_preauth_groups, random_key,
+    s2k_params, strip_db_args,
 };
 
 use krb5_types::PrincipalName;